requirements.txt
# PROJECT OVERSEER - Attack Surface Mapper
# Passive Reconnaissance Tool Requirements

requests>=2.31.0
pandas>=2.0.0
folium>=0.15.0
dnspython>=2.4.0
tqdm>=4.66.0
rich>=13.7.0
